[INDDEX] Simplify CouchDB management feature
